+++ current_version.ts
@@ -1,2 +1,2 @@
-export const CURR_VER_DEV = '1.155';
-export const CURR_VER_PROD = '1.019';
+export const CURR_VER_DEV = '1.153';
+export const CURR_VER_PROD = '1.017';