+++ src/app/utils/globals.ts
@@ -386,7 +386,6 @@ export class Globals {
       this.userFullname = TEMP;
     }
     TEMP = el.nativeElement.getAttribute('preChatForm');
-    console.log('preChatForm', TEMP);
     if (TEMP !== null) {
       this.preChatForm = (TEMP === true) ? true : false;
     }
@@ -513,11 +512,9 @@ export class Globals {
       this.userFullname = TEMP;
     }
     TEMP = window['tiledeskSettings']['preChatForm'];
-    console.log('tiledeskSettings preChatForm', TEMP);
     if (TEMP !== undefined) {
       this.preChatForm = (TEMP === false) ? false : true;
     }
-    console.log('tiledeskSettings preChatForm', this.preChatForm);
 
     TEMP = window['tiledeskSettings']['isOpen'];
     if (TEMP !== undefined) {